Full Name, Picture, and Biography
getting full name, profile picture and biography should work just fine. when you click on the profile picture, it should allow you to choose a picture, and that picture will be saved in the firebase storage. and it will load that picture on the profile pic icon when you sign back in.